# hyperliquid-paper-trader/swagger.yaml
openapi: 3.0.0
info:
  title: Hyperliquid Paper Trader API
  description: API for managing strategies and placing paper trades on Hyperliquid.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Server

paths:
  /strategies:
    post:
      summary: Initialize a new Strategy Wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - capital
              properties:
                id:
                  type: string
                  example: "SWING_01"
                capital:
                  type: number
                  example: 1000
      responses:
        200:
          description: Strategy created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

  /strategies/risk:
    post:
      summary: Update Strategy Risk Settings (Max Daily Loss)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - max_daily_loss
              properties:
                id:
                  type: string
                  example: "SCALP_01"
                max_daily_loss:
                  type: number
                  description: Max daily loss percentage (e.g. 5 for 5%)
                  example: 5
      responses:
        200:
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /order:
    post:
      summary: Place a Paper Trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - strategyId
                - symbol
                - side
                - size
              properties:
                strategyId:
                  type: string
                  example: "SWING_01"
                symbol:
                  type: string
                  example: "BTC"
                side:
                  type: string
                  enum: [LONG, SHORT]
                  example: "LONG"
                size:
                  type: number
                  description: Position Size (Notional Value in USD)
                  example: 500
                sl:
                  type: number
                  description: Stop Loss Price
                  example: 95000
                tp:
                  type: number
                  description: Take Profit Price
                  example: 105000
      responses:
        200:
          description: Order placed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  position:
                    type: object

  /strategies/{id}:
    get:
      summary: Get Strategy Status & History
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Strategy Details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  current_balance:
                    type: number
                  positions:
                    type: array
                    description: Active Positions
                    items:
                      type: object
                  recent_history:
                    type: array
                    description: Closed Positions History
                    items:
                      type: object

    get:
      summary: Get All Strategies
      responses:
        200:
          description: List of all strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    initial_capital:
                      type: number
                    current_balance:
                      type: number
                    pnl:
                      type: string
                    roi:
                      type: string

  /position/close:
    post:
      summary: Manually Close a Position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  example: "123"
      responses:
        200:
          description: Position closed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /position/update:
    post:
      summary: Update TP/SL for a Position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  example: "123"
                tp:
                  type: number
                  description: Take Profit Price
                  example: 105000
                sl:
                  type: number
                  description: Stop Loss Price
                  example: 95000
      responses:
        200:
          description: Position updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  position:
                    type: object

  /position/close-all:
    post:
      summary: Close All Position for a Strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - strategyId
              properties:
                strategyId:
                  type: string
                  example: "SCALP_01"
      responses:
        200:
          description: All positions closed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer